+++ src/main/java/net/frozenblock/wilderwild/registry/WilderWildRegistries.java
@@ -24,7 +24,7 @@ import net.fabricmc.fabric.api.event.registry.RegistryAttribute;
 import net.frozenblock.wilderwild.WWConstants;
 import net.frozenblock.wilderwild.entity.variant.ButterflyVariant;
 import net.frozenblock.wilderwild.entity.variant.firefly.FireflyColor;
-import net.frozenblock.wilderwild.entity.variant.JellyfishVariant;
+import net.frozenblock.wilderwild.entity.variant.jellyfish.JellyfishVariant;
 import net.minecraft.core.MappedRegistry;
 import net.minecraft.core.Registry;
 import net.minecraft.resources.ResourceKey;
@@ -32,12 +32,9 @@ import net.minecraft.resources.ResourceKey;
 public final class WilderWildRegistries {
 	public static final ResourceKey<Registry<FireflyColor>> FIREFLY_COLOR = ResourceKey.createRegistryKey(WWConstants.id("firefly_color"));
 	public static final ResourceKey<Registry<ButterflyVariant>> BUTTERFLY_VARIANT_REGISTRY = ResourceKey.createRegistryKey(WWConstants.id("butterfly_variant"));
-	public static final ResourceKey<Registry<JellyfishVariant>> JELLYFISH_VARIANT_REGISTRY = ResourceKey.createRegistryKey(WWConstants.id("jellyfish_variant"));
+	public static final ResourceKey<Registry<JellyfishVariant>> JELLYFISH_VARIANT = ResourceKey.createRegistryKey(WWConstants.id("jellyfish_variant"));
 
 	public static final MappedRegistry<ButterflyVariant> BUTTERFLY_VARIANT = FabricRegistryBuilder.createSimple(BUTTERFLY_VARIANT_REGISTRY)
-		.attribute(RegistryAttribute.SYNCED)
-		.buildAndRegister();
-	public static final MappedRegistry<JellyfishVariant> JELLYFISH_VARIANT = FabricRegistryBuilder.createSimple(JELLYFISH_VARIANT_REGISTRY)
 		.attribute(RegistryAttribute.SYNCED)
 		.buildAndRegister();
 
@@ -47,8 +44,8 @@ public final class WilderWildRegistries {
 
 	public static void initRegistry() {
 		DynamicRegistries.registerSynced(FIREFLY_COLOR, FireflyColor.DIRECT_CODEC);
+		DynamicRegistries.registerSynced(JELLYFISH_VARIANT, JellyfishVariant.DIRECT_CODEC);
 
 		ButterflyVariant.init();
-		JellyfishVariant.init();
 	}
 }